# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mana Rythu API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: verified
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/verify:
    patch:
      operationId: verifyUser
      tags: [admin]
      summary: Toggle user verification (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyUserInput"
      responses:
        "200":
          description: Verification updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/rate:
    post:
      operationId: rateUser
      tags: [users]
      summary: Rate a farmer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "200":
          description: Rating submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [farmer, buyer, admin]
        verified:
          type: boolean
        rating:
          type: ["number", "null"]
        ratingCount:
          type: integer
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, email, role, verified, ratingCount, createdAt]
...
    VerifyUserInput:
      type: object
      properties:
        verified:
          type: boolean
      required: [verified]
    RatingInput:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
      required: [rating]
...