# lib/api-zod/src/generated/types/ratingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mana Rythu API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RatingInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
}